pyproject.toml
[project]
name = "aim_mcp_demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.3.27",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "mcp[cli]>=1.6.0",
    "numpy>=2.2.4",
    "openai>=1.72.0",
    "python-dotenv>=1.1.0",
    "requests>=2.31.0",
    "tavily-python>=0.5.4",
]
